+++ Utilities/otbossim/src/ossim/imaging/ossimHsvGridRemapEngine.cpp
@@ -14,7 +14,7 @@
 // LIMITATIONS: None.
 //
 //*****************************************************************************
-//  $Id: ossimHsvGridRemapEngine.cpp 9094 2006-06-13 19:12:40Z dburken $
+//  $Id: ossimHsvGridRemapEngine.cpp 15833 2009-10-29 01:41:53Z eshirschorn $
 
 #include <ossim/imaging/ossimHsvGridRemapEngine.h>
 
@@ -162,7 +162,7 @@ void ossimHsvGridRemapEngine::assignRemapValues (
    // Declare a 2D array that will contain all of the contributing sources'
    // HSV mean values. Also declare the accumulator target vector.
    //***
-   int num_contributors = sources_list.size();
+   int num_contributors = (int)sources_list.size();
    double** contributor_pixel = new double* [num_contributors];
    for (i=0; i<num_contributors; i++)
       contributor_pixel[i] = new double[3];
